loan_management/requirements.txt
Django>=4.2
djangorestframework
pandas
openpyxl
psycopg2-binary
celery
redis
python-decouple